requirements.txt
requests>=2.31.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
pylint>=3.0.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0 